+++ config/set/php-cs-fixer.php
@@ -13,6 +13,7 @@ use PhpCsFixer\Fixer\Operator\ConcatSpaceFixer;
 use PhpCsFixer\Fixer\Operator\IncrementStyleFixer;
 use PhpCsFixer\Fixer\Operator\LogicalOperatorsFixer;
 use PhpCsFixer\Fixer\Operator\NotOperatorWithSuccessorSpaceFixer;
+use PhpCsFixer\Fixer\Operator\UnaryOperatorSpacesFixer;
 use PhpCsFixer\Fixer\Phpdoc\NoSuperfluousPhpdocTagsFixer;
 use PhpCsFixer\Fixer\Phpdoc\PhpdocAlignFixer;
 use PhpCsFixer\Fixer\Phpdoc\PhpdocNoAliasTagFixer;
@@ -25,10 +26,13 @@ use PhpCsFixer\Fixer\PhpUnit\PhpUnitTestClassRequiresCoversFixer;
 use PhpCsFixer\Fixer\ReturnNotation\SimplifiedNullReturnFixer;
 use PhpCsFixer\Fixer\Semicolon\MultilineWhitespaceBeforeSemicolonsFixer;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
-use Symplify\EasyCodingStandard\Configuration\Option;
+use Symplify\EasyCodingStandard\ValueObject\Option;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
-    $containerConfigurator->import(__DIR__ . '/php-cs-fixer/php-cs-fixer.php', null, true);
+    $containerConfigurator->import(__DIR__ . '/../../vendor/symplify/easy-coding-standard/config/set/php-cs-fixer.php', null, true);
+
+    $containerConfigurator->import(__DIR__ . '/../../../../symplify/easy-coding-standard/config/set/php-cs-fixer.php', null, true);
+
     $services = $containerConfigurator->services();
     $services->set(ClassAttributesSeparationFixer::class);
     $services->set(MultilineWhitespaceBeforeSemicolonsFixer::class);
@@ -45,23 +49,76 @@ return static function (ContainerConfigurator $containerConfigurator): void {
             ]
         );
     $services->set(ConcatSpaceFixer::class)
-        ->call('configure', [['spacing' => 'one']]);
+        ->call(
+            'configure',
+            [
+                [
+                    'spacing' => 'one',
+                ],
+            ]
+        );
     $services->set(PhpdocAlignFixer::class)
-        ->call('configure', [['align' => 'left']]);
+        ->call(
+            'configure',
+            [
+                [
+                    'align' => 'left',
+                ],
+            ]
+        );
     $services->set(ArraySyntaxFixer::class)
-        ->call('configure', [['syntax' => 'short']]);
+        ->call(
+            'configure',
+            [
+                [
+                    'syntax' => 'short',
+                ],
+            ]
+        );
     $services->set(VisibilityRequiredFixer::class)
-        ->call('configure', [['elements' => ['property', 'method', 'const']]]);
+        ->call(
+            'configure',
+            [
+                [
+                    'elements' => ['property', 'method', 'const'],
+                ],
+            ]
+        );
 
     $services->set(OrderedClassElementsFixer::class)
-        ->call('configure', [['order' => ['use_trait', 'constant_public', 'constant_protected', 'constant_private']]]);
+        ->call(
+            'configure',
+            [
+                [
+                    'order' => ['use_trait', 'constant_public', 'constant_protected', 'constant_private'],
+                ],
+            ]
+        );
 
     $services->set(IncrementStyleFixer::class)
-        ->call('configure', [['style' => 'post']]);
+        ->call(
+            'configure',
+            [
+                [
+                    'style' => 'post',
+                ],
+            ]
+        );
+    $services->set(PhpdocTypesOrderFixer::class)
+        ->call(
+            'configure',
+            [
+                [
+                    'null_adjustment' => 'always_last',
+                    'sort_algorithm' => 'none',
+                ],
+            ]
+        );
     $parameters = $containerConfigurator->parameters();
     $parameters->set(
         Option::SKIP,
         [
+            UnaryOperatorSpacesFixer::class => null,
             SingleLineThrowFixer::class => null,
             PhpdocSummaryFixer::class => null,
             PhpdocToCommentFixer::class => null,