+++ config/set/php_codesniffer.php
@@ -101,7 +101,7 @@ use PHP_CodeSniffer\Standards\Squiz\Sniffs\WhiteSpace\ScopeKeywordSpacingSniff;
 use PHP_CodeSniffer\Standards\Squiz\Sniffs\WhiteSpace\SuperfluousWhitespaceSniff;
 use PHP_CodeSniffer\Standards\Zend\Sniffs\Debug\CodeAnalyzerSniff;
 use PHP_CodeSniffer\Standards\Zend\Sniffs\Files\ClosingTagSniff;
-use Symplify\EasyCodingStandard\Configuration\Option;
+use Symplify\EasyCodingStandard\ValueObject\Option;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
     $services = $containerConfigurator->services();
@@ -243,6 +243,7 @@ return static function (ContainerConfigurator $containerConfigurator): void {
             UseDeclarationSniff::class . '.UseAfterBrace' => null,
             AnonClassDeclarationSniff::class . '.SpaceAfterKeyword' => null,
             ControlStructureSpacingSniff::class . '.LineAfterClose' => null,
+            PostStatementCommentSniff::class . '.Found' => null,
         ]
     );
 };