+++ correct/TestClass.php
@@ -119,4 +119,31 @@ final class TestClass implements TestableContract
     {
         return PHP_EOL;
     }
+
+    public function testSingleArray()
+    {
+        return ['a', 'b']; // post statement comment
+    }
+
+    public function testScopeClosingBrace()
+    {
+        return [
+            'eventCrowd' => function ($query) {
+                return $query->with('rule');
+            },
+            'eventAuction' => function ($query) {
+                return $query->with('rule');
+            },
+            'logs' => function ($query) {
+                return $query->with(
+                    [
+                        'admin' => function ($query) {
+                            return $query->select('id', 'name');
+                        },
+                    ]
+                );
+            },
+            'images',
+        ];
+    }
 }