+++ CHANGELOG.md
@@ -13,6 +13,58 @@ PRs and issues are linked, so you can find more about it. Thanks to [ChangelogLi
 
 ### Changed
 
+- [#45] Re-enable ArrayListItemNewlineFixer
+
+## [2.6.0] - 2020-09-30
+
+### Changed
+
+- [#44] Disable ArrayListItemNewlineFixer
+- [#43] Use constant instead of string for changelog-linker option
+
+### Deprecated
+
+- [#42] Remove deprecated usage of Symplify\EasyCodingStandard\Configuration\Option
+
+### Removed
+
+- [#41] Remove php-cs-fixer rule sets
+
+## [2.5.4] - 2020-09-27
+
+### Changed
+
+- [#40] Disable PostStatementCommentSniff.Found
+
+## [2.5.3] - 2020-09-25
+
+### Changed
+
+- [#39] Bump rector/rector from 0.8.0 to 0.8.10
+
+## [2.5.2] - 2020-09-22
+
+### Fixed
+
+- [#38] Fix phpstan/phpstan ^0.12.43 requires unsupported nikic/php-parser 4.10.0
+
+## [2.5.1] - 2020-09-16
+
+### Changed
+
+- [#36] Skip UnaryOperatorSpacesFixer
+
+## [2.5.0] - 2020-09-02
+
+### Changed
+
+- [#35] YAML to PHP
+- [#34] Update dependency rector/rector to ^0.8.0
+
+## [2.4.0] - 2020-08-31
+
+### Changed
+
 - [#33] Update slevomat/coding-standard from dev-master to ^6.4
 
 ## [2.3.0] - 2020-08-21
@@ -154,3 +206,21 @@ PRs and issues are linked, so you can find more about it. Thanks to [ChangelogLi
 [2.2.2]: https://github.com/zingimmick/coding-standard/compare/2.2.1...2.2.2
 [#33]: https://github.com/zingimmick/coding-standard/pull/33
 [2.3.0]: https://github.com/zingimmick/coding-standard/compare/2.2.2...2.3.0
+[#35]: https://github.com/zingimmick/coding-standard/pull/35
+[#34]: https://github.com/zingimmick/coding-standard/pull/34
+[2.4.0]: https://github.com/zingimmick/coding-standard/compare/2.3.0...2.4.0
+[#36]: https://github.com/zingimmick/coding-standard/pull/36
+[2.5.0]: https://github.com/zingimmick/coding-standard/compare/2.4.0...2.5.0
+[#38]: https://github.com/zingimmick/coding-standard/pull/38
+[2.5.1]: https://github.com/zingimmick/coding-standard/compare/2.5.0...2.5.1
+[#39]: https://github.com/zingimmick/coding-standard/pull/39
+[2.5.2]: https://github.com/zingimmick/coding-standard/compare/2.5.1...2.5.2
+[#40]: https://github.com/zingimmick/coding-standard/pull/40
+[2.5.3]: https://github.com/zingimmick/coding-standard/compare/2.5.2...2.5.3
+[#44]: https://github.com/zingimmick/coding-standard/pull/44
+[#43]: https://github.com/zingimmick/coding-standard/pull/43
+[#42]: https://github.com/zingimmick/coding-standard/pull/42
+[#41]: https://github.com/zingimmick/coding-standard/pull/41
+[2.5.4]: https://github.com/zingimmick/coding-standard/compare/2.5.3...2.5.4
+[#45]: https://github.com/zingimmick/coding-standard/pull/45
+[2.6.0]: https://github.com/zingimmick/coding-standard/compare/2.5.4...2.6.0